+++ example/index.ts
@@ -18,158 +18,17 @@ const user = t.Object({
 	})
 })
 
-export const app = new Elysia()
-	.use(
-		openapi({
-			provider: 'scalar',
-			mapJsonSchema: {
-				zod: z.toJSONSchema,
-				effect: JSONSchema.make
-			},
-			documentation: {
-				info: {
-					title: 'Elysia Scalar',
-					version: '1.3.1a'
-				},
-				tags: [
-					{
-						name: 'Test',
-						description: 'Hello'
-					}
-				],
-				components: {
-					securitySchemes: {
-						bearer: {
-							type: 'http',
-							scheme: 'bearer'
-						},
-						cookie: {
-							type: 'apiKey',
-							in: 'cookie',
-							name: 'session_id'
-						}
-					}
-				}
-			}
-		})
-	)
-	.model({ schema, schema2, user })
-	.model({
-		idParam: t.Object({
-			id: t.Union([
-				t.String({ format: 'uuid' }),
-				t.Number({ minimum: 1, maximum: Number.MAX_SAFE_INTEGER })
-			]),
-			id2: t.String()
-		}),
-		response200: t.Object({
-			message: t.String(),
-			content: t.Array(t.Object({
-				id: t.Union([t.String(), t.Number()])
-			}))
-		})
+const model = new Elysia().model(
+	'body',
+	t.Object({
+		name: t.Literal('Lilith')
 	})
-	.get('/test/:id/:id2', ({ params: { id } }) => ({
-		message: 'ok',
-		content: [{ id }]
-	}), {
-		params: 'idParam',
-		response: 'response200'
+)
+
+const app = new Elysia()
+	.use(openapi())
+	.use(model)
+	.post('/user', () => 'hello', {
+		body: 'body'
 	})
-	.get(
-		'/',
-		{ test: 'hello' as const },
-		{
-			response: {
-				200: t.Object({
-					test: t.Literal('hello')
-				}),
-				204: withHeaders(
-					t.Void({
-						title: 'Thing',
-						description: 'Void response'
-					}),
-					{
-						'X-Custom-Header': t.Literal('Elysia')
-					}
-				)
-			}
-		}
-	)
-	.post(
-		'/json',
-		({ body }) => ({
-			test: 'hello'
-		}),
-		{
-			parse: ['json', 'formdata'],
-			body: 'schema',
-			response: {
-				200: t.Object({
-					test: t.Literal('hello')
-				}),
-				400: z.object({
-					a: z.string(),
-					b: z.literal('a')
-				}),
-				401: Schema.standardSchemaV1(
-					Schema.Struct({
-						a: Schema.Literal('hi')
-					})
-				)
-			}
-		}
-	)
-	.post(
-		'/json/:id',
-		({ body, params: { id }, query: { name, email, birthday } }) => ({
-			...body,
-			id,
-			name,
-			email,
-			birthday
-		}),
-		{
-			params: t.Object({
-				id: t.String()
-			}),
-			query: t.Object({
-				name: t.String(),
-				email: t.String({
-					description: 'sample email description',
-					format: 'email'
-				}),
-				birthday: t.String({
-					description: 'sample birthday description',
-					pattern: '\\d{4}-\\d{2}-\\d{2}',
-					minLength: 10,
-					maxLength: 10
-				})
-			}),
-			body: t.Object({
-				username: t.String(),
-				password: t.String()
-			}),
-			response: t.Object(
-				{
-					username: t.String(),
-					password: t.String(),
-					id: t.String(),
-					name: t.String(),
-					email: t.String({
-						description: 'sample email description',
-						format: 'email'
-					}),
-					birthday: t.String({
-						description: 'sample birthday description',
-						pattern: '\\d{4}-\\d{2}-\\d{2}',
-						minLength: 10,
-						maxLength: 10
-					})
-				},
-				{ description: 'sample description 3' }
-			)
-		}
-	)
-	.get('/id/:id/name/:name', () => {})
 	.listen(3000)